Restructured event handler code
