+++ test/bot/test_greet_member.py
@@ -6,10 +6,10 @@ from discord import Embed
 
 # noinspection PyUnresolvedReferences
 import test.context
-from src.event.on_member_join import on_member_join
+from event.on_member_join.greet_member import greet_member
 
 
-class TestOnMemberJoin(aiounittest.AsyncTestCase):
+class TestGreetMember(aiounittest.AsyncTestCase):
 
     @mock.patch('src.bot.BotClient', autospec=True)
     def setUp(self, bot):
@@ -22,7 +22,7 @@ class TestOnMemberJoin(aiounittest.AsyncTestCase):
         guild.system_channel.send = mock.Mock(futurized(None))
         member.guild = guild
         # could also do `self.bot.on_member_join(member)` here but I want to stay consistent across tests.
-        await on_member_join(self.bot)(member)
+        await greet_member(self.bot)(member)
         guild.system_channel.send.assert_called_once()
         # first argument of first call
         embed = guild.system_channel.send.call_args[0][0]
@@ -35,4 +35,4 @@ class TestOnMemberJoin(aiounittest.AsyncTestCase):
         member.guild = guild
         # TODO when you figured out testing logging, add here test for warning log
         with self.assertRaises(RuntimeWarning):
-            await on_member_join(self.bot)(member)
+            await greet_member(self.bot)(member)
