+++ test/bot/test_lecture_role_assignment.py
@@ -6,43 +6,43 @@ from aiounittest import futurized
 # noinspection PyUnresolvedReferences
 import test.context
 from src.const import WHITE_CHECK_MARK
-from src.event.on_raw_reaction_remove import on_raw_reaction_remove
+from src.event.on_raw_reaction_add.lecture_role_assignment import lecture_role_assignment
 
 
-class TestOnRawReactionRemove(aiounittest.AsyncTestCase):
+class TestLectureRoleAssignment(aiounittest.AsyncTestCase):
 
     @mock.patch('discord.Role', autospec=True)
     @mock.patch('discord.Member', autospec=True)
     @mock.patch('discord.Emoji', autospec=True)
-    @mock.patch('discord.Guild', autospec=True)
     @mock.patch('discord.RawReactionActionEvent', autospec=True)
     @mock.patch('src.util.lecture.Lecture', autospec=True)
     @mock.patch('src.bot.BotClient', autospec=True)
-    def setUp(self, bot, lecture, reaction, guild, emoji, member, role):
+    def setUp(self, bot, lecture, reaction, emoji, member, role):
+        """Since we use `autospec` to make sure that we use the actual bot API, we run into the problem
+        that accessing instance attributes now leads to errors because "`autospec` can't know about
+        any dynamically created attributes and restricts the api to visible attributes".
+        ( See https://docs.python.org/3/library/unittest.mock.html#auto-speccing )
+        We therefore mock the instance variables we will need when testing the SUT."""
         bot.logger = mock.Mock()
         bot.user.id = '00000'
         self.bot = bot
+        # setup the lecture mock we will receive when calling bot#get_lecture_of_message_id
         # when calling lecture.role, we want to get the "role id"
         lecture.role = '1234'
         lecture.embed_title = 'title'  # not needed to test behaviour but to fix unimportant AttributeError during test
         # bot#get_lecture_of_message_id should return the mocked lecture
         self.bot.get_lecture_of_message_id.return_value = lecture
-        """member#remove_roles should always be awaitable even though we are not always expecting that this function will
+        """member#add_roles should always be awaitable even though we are not always expecting that this function will
         be called. The reason for this that we don't want to couple our test to tightly with the code; expecting more
         from the SUT (system under test) than we are testing.
         We assert that it was or was not alled but don't want to throw an error during testing just because it
         is not awaitable. The actual reason to fail the test should be that it was or was not called!
         """
-        member.remove_roles = mock.Mock(futurized(None))
+        member.add_roles = mock.Mock(futurized(None))
         # member.guild#get_role should return the mocked role
         member.guild.get_role.return_value = role
-        # since in event "REACTION_REMOVE", the member is not added to the reaction, we get the member through the guild
-        # See https://discordpy.readthedocs.io/en/latest/api.html#discord.RawReactionActionEvent
-        guild.get_member.return_value = member
-        # bot#get_guild will be called with the guild id of reaction and should return our mocked guild
-        self.bot.get_guild.return_value = guild
-        reaction.guild_id = 99999
-        # add mock to reaction
+        # add mocks to reaction
+        reaction.member = member
         reaction.emoji = emoji
         # add mocks to instance so we can access them in the test cases
         self.reaction = reaction
@@ -51,40 +51,39 @@ class TestOnRawReactionRemove(aiounittest.AsyncTestCase):
         self.role = role
         self.lecture = lecture
 
-    async def test_on_raw_reaction_remove_removes_role_when_removed_white_check_mark_reaction_from_lecture_embed(self):
+    async def test_on_raw_reaction_add_adds_role_when_reacted_with_white_check_mark_on_lecture_embed(self):
         reaction, emoji, member, role = self.reaction, self.emoji, self.member, self.role
-        # user removed reaction WHITE_CHECK_MARK
+        # user reacted with WHITE_CHECK_MARK
         emoji.name = WHITE_CHECK_MARK
         # user reacted to message with message id 5678
         reaction.message_id = '5678'
         # it was an actual user which reacted, not our bot
         reaction.user_id = '11111'  # not equal to bot.user.id
 
-        await on_raw_reaction_remove(self.bot)(reaction)
+        await lecture_role_assignment(self.bot)(reaction)
 
         # assert that we tried to find the lecture via the message id
         self.bot.get_lecture_of_message_id.assert_called_once_with('5678')
         # assert that we got the role from the guild with its id as integer
         member.guild.get_role.assert_called_once_with(1234)
-        # assert that we removed the role from the member
-        member.remove_roles.assert_called_once_with(role)
+        # assert that we added the role to the member
+        member.add_roles.assert_called_once_with(role)
 
-    async def test_on_raw_reaction_remove_does_not_remove_role_when_removed_reaction_from_lecture_embed_was_not_white_check_mark(
-            self):
-        reaction, emoji, member = self.reaction, self.emoji, self.member
-        # user removed another reaction
+    async def test_on_raw_reaction_add_does_not_add_role_when_not_reacted_with_white_check_mark_on_lecture_embed(self):
+        reaction, emoji, member, role = self.reaction, self.emoji, self.member, self.role
+        # user reacted with something else than WHITE_CHECK_MARK
         emoji.name = WHITE_CHECK_MARK + "xx"  # TODO Create another actual emoji unicode character for usage here
         # user reacted to message with message id 5678
         reaction.message_id = '5678'
         # it was an actual user which reacted, not our bot
-        reaction.user_id = '1111'  # not equal to bot.user.id
+        reaction.user_id = '11111'  # not equal to bot.user.id
 
-        await on_raw_reaction_remove(self.bot)(reaction)
+        await lecture_role_assignment(self.bot)(reaction)
 
-        # assert that we did not call member#remove_roles
+        # assert that we did not call member#add_role
         member.add_roles.assert_not_called()
 
-    async def test_on_raw_reaction_remove_ignores_reaction_from_bot(self):
+    async def test_on_raw_reaction_add_ignores_reaction_from_bot(self):
         reaction, emoji, member = self.reaction, self.emoji, self.member
         # the reaction was from the bot itself
         reaction.user_id = self.bot.user.id
@@ -93,7 +92,7 @@ class TestOnRawReactionRemove(aiounittest.AsyncTestCase):
         # and to message with id 5678
         reaction.message_id = '5678'
 
-        await on_raw_reaction_remove(self.bot)(reaction)
+        await lecture_role_assignment(self.bot)(reaction)
 
         # assert that none of the following methods were called:
         self.bot.get_lecture_of_message_id.assert_not_called()
